доделана модуляция + сделаны миниатюры
