+++ index.html
@@ -9,7 +9,7 @@
   <title>Кекстаграм</title>
 
   <scpirt src="js/functions.js"></scpirt>
-  <scpirt src="js/main.js"></scpirt>
+  <scpirt type="module" src="js/main.js"></scpirt>
 </head>
 
 <body>
